requirements.txt
Flask
Flask-Bcrypt
Flask-WTF
Flask-Login
Flask-SQLAlchemy
WTForms
Pillow
gunicorn
email_validator